Split odt.rs into multiple files
Also created a new folder for it
