+++ server/src/parsers/odt/table.rs
@@ -1,0 +1,720 @@
+use super::*;
+
+impl ODTParser {
+    /// Helper for handle_element_start() to respond to tags with "table" prefix
+    pub fn handle_element_start_table(&mut self, local_name: &str, attributes: Attributes) {
+        match local_name {
+            "table" => {
+                self.document_hierarchy
+                    .push(table_begin(attributes, &self.auto_styles));
+                self.table_column_default_style_names.push(Vec::new());
+                self.table_column_number.push(0);
+                self.table_row_default_style_names.push(Vec::new());
+                self.table_row_number.push(0);
+            }
+            "table-row" => {
+                if self.table_row_default_style_names.is_empty() {
+                    return;
+                }
+                let (row, default_cell_style_name) = table_row_begin(attributes, &self.auto_styles);
+                self.document_hierarchy.push(row);
+                let table_row_default_style_names =
+                    self.table_row_default_style_names.last_mut().unwrap();
+                if let Some(default_cell_style_name) = default_cell_style_name {
+                    table_row_default_style_names.push(default_cell_style_name.clone());
+                } else {
+                    table_row_default_style_names.push("".to_string());
+                }
+                let len = self.table_column_number.len();
+                self.table_column_number[len - 1] = 0;
+            }
+            "table-cell" => {
+                // This should not be empty, just check this since the other three will be set up together with it
+                if self.table_row_default_style_names.is_empty() {
+                    return;
+                }
+                let mut default_style_name = String::new();
+                let row_default = &self.table_row_default_style_names.last().unwrap()
+                    [*self.table_row_number.last().unwrap() as usize];
+                let column_default = &self.table_column_default_style_names.last().unwrap()
+                    [*self.table_column_number.last().unwrap() as usize];
+                if row_default != "" {
+                    default_style_name = row_default.to_string();
+                } else if column_default != "" {
+                    default_style_name = column_default.to_string();
+                }
+                self.document_hierarchy.push(table_cell_begin(
+                    attributes,
+                    &self.auto_styles,
+                    default_style_name,
+                ));
+            }
+            _ => (),
+        }
+    }
+
+    /// Helper for handle_element_end() to respond to tags with "table" prefix
+    pub fn handle_element_end_table(&mut self, local_name: &str) {
+        match local_name {
+            "table" => {
+                let child = self.document_hierarchy.pop().unwrap();
+                if self.document_hierarchy.is_empty() {
+                    self.document_root.children.push(Node::Element(child));
+                } else {
+                    self.document_hierarchy
+                        .last_mut()
+                        .unwrap()
+                        .children
+                        .push(Node::Element(child));
+                }
+                self.table_column_default_style_names.pop();
+                self.table_column_number.pop();
+                self.table_row_default_style_names.pop();
+                self.table_row_number.pop();
+            }
+            "table-row" => {
+                let mut child = self.document_hierarchy.pop().unwrap();
+                let mut repeat = child
+                    .attributes
+                    .remove("_repeat")
+                    .unwrap_or_else(|| "1".to_string())
+                    .parse::<u32>()
+                    .unwrap_or(1);
+                let len = self.table_row_number.len();
+                self.table_row_number[len - 1] += repeat;
+                while repeat != 0 {
+                    if self.document_hierarchy.is_empty() {
+                        self.document_root
+                            .children
+                            .push(Node::Element(child.clone()));
+                    } else {
+                        self.document_hierarchy
+                            .last_mut()
+                            .unwrap()
+                            .children
+                            .push(Node::Element(child.clone()));
+                    }
+                    repeat -= 1;
+                }
+            }
+            "table-cell" => {
+                let mut child = self.document_hierarchy.pop().unwrap();
+                let mut repeat = child
+                    .attributes
+                    .remove("_repeat")
+                    .unwrap_or_else(|| "1".to_string())
+                    .parse::<u32>()
+                    .unwrap_or(1);
+                let len = self.table_column_number.len();
+                self.table_column_number[len - 1] += repeat;
+                while repeat != 0 {
+                    if self.document_hierarchy.is_empty() {
+                        self.document_root
+                            .children
+                            .push(Node::Element(child.clone()));
+                    } else {
+                        self.document_hierarchy
+                            .last_mut()
+                            .unwrap()
+                            .children
+                            .push(Node::Element(child.clone()));
+                    }
+                    repeat -= 1;
+                }
+            }
+            _ => (),
+        }
+    }
+}
+
+enum TableAlign {
+    Center,
+    Left,
+    Right,
+    Margins,
+}
+
+/// Takes the set of attributes of a style:table-properties tag in the ODT's content.xml,
+/// and creates a map of CSS properties based on the attributes
+pub fn table_properties_begin(attributes: Attributes) -> HashMap<String, String> {
+    let mut map: HashMap<String, String> = HashMap::new();
+    let mut table_alignment = TableAlign::Margins;
+    let mut margin_left = "0cm".to_string();
+    let mut margin_right = "0cm".to_string();
+    for i in attributes {
+        if let Ok(i) = i {
+            let name = std::str::from_utf8(i.key).unwrap_or(":");
+            let (prefix, local_name) = name.split_at(name.find(':').unwrap_or(0));
+            let local_name = &local_name[1..];
+            let value = std::str::from_utf8(
+                &i.unescaped_value()
+                    .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+            )
+            .unwrap_or("what")
+            .to_string();
+            if prefix == "fo" {
+                let (margin_left_option, margin_right_option) =
+                    table_properties_begin_fo(local_name, value, &mut map);
+                if let Some(margin_left_option) = margin_left_option {
+                    margin_left = margin_left_option;
+                }
+                if let Some(margin_right_option) = margin_right_option {
+                    margin_right = margin_right_option;
+                }
+            } else if prefix == "style" {
+                table_properties_begin_style(local_name, value, &mut map);
+            } else if prefix == "table" {
+                if let Some(table_alignment_option) =
+                    table_properties_begin_table(local_name, value, &mut map)
+                {
+                    table_alignment = table_alignment_option;
+                }
+            }
+        }
+    }
+    match table_alignment {
+        // The specification says to ignore the right margin if it is aligned left
+        TableAlign::Left => {
+            map.insert("marginRight".to_string(), "unset".to_string());
+        }
+        // Similarly for align right
+        TableAlign::Right => {
+            map.insert("marginLeft".to_string(), "unset".to_string());
+        }
+        // For center alignment we are supposed to ignore both side margins, while in CSS centering is done by making both of them auto
+        TableAlign::Center => {
+            map.insert("marginLeft".to_string(), "auto".to_string());
+            map.insert("marginRight".to_string(), "auto".to_string());
+        }
+        TableAlign::Margins => {
+            map.insert(
+                "width".to_string(),
+                format!("calc(100% - {} - {})", margin_left, margin_right),
+            );
+        }
+    }
+    map
+}
+
+/// Helper for table_properties_begin() for attributes with "fo" prefix,
+/// returns the values of the left and right margin if the attribute specifies either
+fn table_properties_begin_fo(
+    local_name: &str,
+    value: String,
+    map: &mut HashMap<String, String>,
+) -> (Option<String>, Option<String>) {
+    match local_name {
+        "background-color" => {
+            map.insert("fontWeight".to_string(), value);
+        }
+        "break-after" => {
+            if value == "auto" || value == "column" || value == "page" {
+                map.insert("breakAfter".to_string(), value);
+            }
+        }
+        "break-before" => {
+            if value == "auto" || value == "column" || value == "page" {
+                map.insert("breakBefore".to_string(), value);
+            }
+        }
+        "margin" => {
+            let margin_split = value.clone();
+            let mut margin_split = margin_split.split(' ');
+            let right = margin_split.nth(1);
+            let left = margin_split.nth(1);
+            let left = left.unwrap_or("0cm").to_string();
+            let right = right.unwrap_or("0cm").to_string();
+            map.insert("margin".to_string(), value);
+            return (Some(left), Some(right));
+        }
+        "margin-top" => {
+            map.insert("marginTop".to_string(), value);
+        }
+        "margin-bottom" => {
+            map.insert("marginLeft".to_string(), value);
+        }
+        "margin-left" => {
+            let left = value.clone();
+            map.insert("marginRight".to_string(), value);
+            return (Some(left), None);
+        }
+        "margin-right" => {
+            let right = value.clone();
+            map.insert("marginBottom".to_string(), value);
+            return (None, Some(right));
+        }
+        _ => (),
+    }
+    (None, None)
+}
+
+/// Helper for table_properties_begin() for attributes with "style" prefix
+fn table_properties_begin_style(
+    local_name: &str,
+    value: String,
+    map: &mut HashMap<String, String>,
+) {
+    match local_name {
+        "rel-width" | "width" => {
+            map.insert("width".to_string(), value);
+        }
+        "shadow" => {
+            map.insert("boxShadow".to_string(), value);
+        }
+        "writing-mode" => {
+            match value.as_str() {
+                // According to the MDN the replacement for "rl" and "rl-tb" is also "horizontal-tb" apparently
+                "lr-tb" | "lr" | "rl" | "rl-tb" => {
+                    map.insert("writingMode".to_string(), "horizontal-tb".to_string());
+                }
+                // MDN says "tb" is supposed to be replaced by "vertical-lr", but the ODT definition says that "tb" is a synonym for "tb-rl"
+                "tb-rl" | "tb" => {
+                    map.insert("writingMode".to_string(), "vertical-rl".to_string());
+                }
+                "tb-lr" => {
+                    map.insert("writingMode".to_string(), "vertical-lr".to_string());
+                }
+                _ => (),
+            }
+        }
+        _ => (),
+    }
+}
+
+/// Helper for table_properties_begin() for attributes with "table" prefix
+fn table_properties_begin_table(
+    local_name: &str,
+    value: String,
+    map: &mut HashMap<String, String>,
+) -> Option<TableAlign> {
+    match local_name {
+        "align" => match value.as_str() {
+            "center" => return Some(TableAlign::Center),
+            "left" => return Some(TableAlign::Left),
+            "right" => return Some(TableAlign::Right),
+            "margins" => return Some(TableAlign::Margins),
+            _ => return None,
+        },
+        "border-model" => match value.as_str() {
+            "collapsing" => {
+                map.insert("borderCollapse".to_string(), "collapse".to_string());
+            }
+            "separating" => {
+                map.insert("borderCollapse".to_string(), "separate".to_string());
+            }
+            _ => (),
+        },
+        "display" => {
+            if value == "false" {
+                map.insert("display".to_string(), "none".to_string());
+            }
+        }
+        _ => (),
+    }
+    None
+}
+
+/// Takes the set of attributes of a style:table-column-properties tag in the ODT's content.xml,
+/// and creates a map of CSS properties based on the attributes
+pub fn table_column_properties_begin(attributes: Attributes) -> HashMap<String, String> {
+    let mut map: HashMap<String, String> = HashMap::new();
+    for i in attributes {
+        if let Ok(i) = i {
+            let name = std::str::from_utf8(i.key).unwrap_or(":");
+            let value = std::str::from_utf8(
+                &i.unescaped_value()
+                    .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+            )
+            .unwrap_or("what")
+            .to_string();
+            match name {
+                "fo:break-after" => {
+                    if value == "auto" || value == "column" || value == "page" {
+                        map.insert("breakAfter".to_string(), value);
+                    }
+                }
+                "fo:break-before" => {
+                    if value == "auto" || value == "column" || value == "page" {
+                        map.insert("breakBefore".to_string(), value);
+                    }
+                }
+                "style:column-width" => {
+                    map.insert("width".to_string(), value);
+                }
+                "style:use-optimal-column-width" => {
+                    if value == "true" {
+                        map.insert("width".to_string(), "auto".to_string());
+                    }
+                }
+                _ => (),
+            }
+        }
+    }
+    map
+}
+
+/// Takes the set of attributes of a style:table-row-properties tag in the ODT's content.xml,
+/// and creates a map of CSS properties based on the attributes
+pub fn table_row_properties_begin(attributes: Attributes) -> HashMap<String, String> {
+    let mut map: HashMap<String, String> = HashMap::new();
+    for i in attributes {
+        if let Ok(i) = i {
+            let name = std::str::from_utf8(i.key).unwrap_or(":");
+            let value = std::str::from_utf8(
+                &i.unescaped_value()
+                    .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+            )
+            .unwrap_or("what")
+            .to_string();
+            match name {
+                "fo:break-after" => {
+                    if value == "auto" || value == "column" || value == "page" {
+                        map.insert("breakAfter".to_string(), value);
+                    }
+                }
+                "fo:break-before" => {
+                    if value == "auto" || value == "column" || value == "page" {
+                        map.insert("breakBefore".to_string(), value);
+                    }
+                }
+                "fo:background-color" => {
+                    map.insert("backgroundColor".to_string(), value);
+                }
+                "style:row-height" => {
+                    map.insert("height".to_string(), value);
+                }
+                "style:use-optimal-row-height" => {
+                    if value == "true" {
+                        map.insert("height".to_string(), "auto".to_string());
+                    }
+                }
+                "style:min-row-height" => {
+                    map.insert("min-height".to_string(), value);
+                }
+                _ => (),
+            }
+        }
+    }
+    map
+}
+
+/// Takes the set of attributes of a style:table-cell-properties tag in the ODT's content.xml,
+/// and creates a map of CSS properties based on the attributes
+pub fn table_cell_properties_begin(attributes: Attributes) -> HashMap<String, String> {
+    let mut map: HashMap<String, String> = HashMap::new();
+    for i in attributes {
+        if let Ok(i) = i {
+            let name = std::str::from_utf8(i.key).unwrap_or(":");
+            let (prefix, local_name) = name.split_at(name.find(':').unwrap_or(0));
+            let local_name = &local_name[1..];
+            let value = std::str::from_utf8(
+                &i.unescaped_value()
+                    .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+            )
+            .unwrap_or("what")
+            .to_string();
+            match prefix {
+                "fo" => table_cell_properties_begin_fo(local_name, value, &mut map),
+                "style" => table_cell_properties_begin_style(local_name, value, &mut map),
+                _ => (),
+            }
+        }
+    }
+    map
+}
+
+/// Helper for table_cell_properties_begin() for attributes with "fo" prefix
+fn table_cell_properties_begin_fo(
+    local_name: &str,
+    value: String,
+    map: &mut HashMap<String, String>,
+) {
+    match local_name {
+        "background-color" => {
+            map.insert("backgroundColor".to_string(), value);
+        }
+        "border" => {
+            map.insert("border".to_string(), value);
+        }
+        "border-left" => {
+            map.insert("borderLeft".to_string(), value);
+        }
+        "border-right" => {
+            map.insert("borderRight".to_string(), value);
+        }
+        "border-top" => {
+            map.insert("borderTop".to_string(), value);
+        }
+        "border-bottom" => {
+            map.insert("borderBottom".to_string(), value);
+        }
+        "padding" => {
+            map.insert("padding".to_string(), value);
+        }
+        "padding-top" => {
+            map.insert("paddingTop".to_string(), value);
+        }
+        "padding-bottom" => {
+            map.insert("paddingBottom".to_string(), value);
+        }
+        "padding-left" => {
+            map.insert("paddingLeft".to_string(), value);
+        }
+        "padding-right" => {
+            map.insert("paddingRight".to_string(), value);
+        }
+        _ => (),
+    }
+}
+
+/// Helper for table_cell_properties_begin() for attributes with "style" prefix
+fn table_cell_properties_begin_style(
+    local_name: &str,
+    value: String,
+    map: &mut HashMap<String, String>,
+) {
+    match local_name {
+        "rotation-angle" => {
+            map.insert("transform".to_string(), format!("rotate({})", value));
+        }
+        "shadow" => {
+            map.insert("boxShadow".to_string(), value);
+        }
+        "writing-mode" => {
+            match value.as_str() {
+                // According to the MDN the replacement for "rl" and "rl-tb" is also "horizontal-tb" apparently
+                "lr-tb" | "lr" | "rl" | "rl-tb" => {
+                    map.insert("writingMode".to_string(), "horizontal-tb".to_string());
+                }
+                // MDN says "tb" is supposed to be replaced by "vertical-lr", but the ODT definition says that "tb" is a synonym for "tb-rl"
+                "tb-rl" | "tb" => {
+                    map.insert("writingMode".to_string(), "vertical-rl".to_string());
+                }
+                "tb-lr" => {
+                    map.insert("writingMode".to_string(), "vertical-lr".to_string());
+                }
+                _ => (),
+            }
+        }
+        "vertical-align" => {
+            if value == "middle" || value == "top" || value == "bottom" {
+                map.insert("verticalAlign".to_string(), value);
+            }
+        }
+        _ => (),
+    }
+}
+
+/// Takes the set of attributes of a table:table tag in the ODT's content.xml
+/// and a reference to the map of automatic style names to the map of CSS properties,
+/// then returns a table element with styles attached
+fn table_begin(
+    attributes: Attributes,
+    auto_styles: &HashMap<String, HashMap<String, String>>,
+) -> Element {
+    let mut style_name = String::new();
+    for i in attributes {
+        if let Ok(i) = i {
+            let name = std::str::from_utf8(i.key).unwrap_or(":");
+            if name == "table:style-name" {
+                style_name = std::str::from_utf8(
+                    &i.unescaped_value()
+                        .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+                )
+                .unwrap_or("")
+                .to_string();
+            }
+        }
+    }
+    let mut element = Element::new("table".to_string());
+    element.styles = auto_styles
+        .get(&style_name)
+        .unwrap_or(&HashMap::new())
+        .clone();
+    // Caption is always the first child of a table in HTML if it exists, and colgroup always comes after if it is there too,
+    // so we add it in in order to have a definitive position as to where the colgroup will come
+    element
+        .children
+        .push(Node::Element(Element::new("caption".to_string())));
+    element
+        .children
+        .push(Node::Element(Element::new("colgroup".to_string())));
+    element
+}
+
+/// Takes the set of attributes of a table:table-column tag in the ODT's content.xml
+/// and a reference to the map of automatic style names to the map of CSS properties,
+/// then returns a table column element with styles attached, the default cell style name (if any)
+/// and how many times it should be repeated (as in the HTML span attribute for columns)
+pub fn table_column_begin(
+    attributes: Attributes,
+    auto_styles: &HashMap<String, HashMap<String, String>>,
+) -> (Element, Option<String>, u32) {
+    let mut style_name = String::new();
+    let mut element = Element::new("col".to_string());
+    let mut repeat = "1".to_string();
+    let mut default_cell_style_name: Option<String> = None;
+    for i in attributes {
+        if let Ok(i) = i {
+            let name = std::str::from_utf8(i.key).unwrap_or(":");
+            if name == "table:style-name" {
+                style_name = std::str::from_utf8(
+                    &i.unescaped_value()
+                        .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+                )
+                .unwrap_or("")
+                .to_string();
+            } else if name == "table:number-columns-repeated" {
+                repeat = std::str::from_utf8(
+                    &i.unescaped_value()
+                        .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+                )
+                .unwrap_or("1")
+                .to_string();
+                element
+                    .attributes
+                    .insert("span".to_string(), repeat.clone());
+            } else if name == "table:default-cell-style-name" {
+                default_cell_style_name = Some(
+                    std::str::from_utf8(
+                        &i.unescaped_value()
+                            .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+                    )
+                    .unwrap_or("1")
+                    .to_string(),
+                );
+            }
+        }
+    }
+    element.styles = auto_styles
+        .get(&style_name)
+        .unwrap_or(&HashMap::new())
+        .clone();
+    let repeat = repeat.parse::<u32>().unwrap_or(1);
+    if let Some(default_cell_style_name) = default_cell_style_name {
+        return (element, Some(default_cell_style_name), repeat);
+    }
+    (element, None, repeat)
+}
+
+/// Takes the set of attributes of a table:table-row tag in the ODT's content.xml
+/// and a reference to the map of automatic style names to the map of CSS properties,
+/// then returns a table row element with styles attached, the default cell style name (if any)
+/// and how many times it should be repeated (stored inside the element as an attribute called "_repeat")
+fn table_row_begin(
+    attributes: Attributes,
+    auto_styles: &HashMap<String, HashMap<String, String>>,
+) -> (Element, Option<String>) {
+    let mut style_name = String::new();
+    let mut element = Element::new("tr".to_string());
+    let mut default_cell_style_name: Option<String> = None;
+    for i in attributes {
+        if let Ok(i) = i {
+            let name = std::str::from_utf8(i.key).unwrap_or(":");
+            if name == "table:style-name" {
+                style_name = std::str::from_utf8(
+                    &i.unescaped_value()
+                        .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+                )
+                .unwrap_or("")
+                .to_string();
+            } else if name == "table:number-rows-repeated" {
+                element.attributes.insert(
+                    "span".to_string(),
+                    std::str::from_utf8(
+                        &i.unescaped_value()
+                            .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+                    )
+                    .unwrap_or("1")
+                    .to_string(),
+                );
+            } else if name == "table:default-cell-style-name" {
+                default_cell_style_name = Some(
+                    std::str::from_utf8(
+                        &i.unescaped_value()
+                            .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+                    )
+                    .unwrap_or("1")
+                    .to_string(),
+                );
+            }
+        }
+    }
+    element.styles = auto_styles
+        .get(&style_name)
+        .unwrap_or(&HashMap::new())
+        .clone();
+    if let Some(default_cell_style_name) = default_cell_style_name {
+        return (element, Some(default_cell_style_name));
+    }
+    (element, None)
+}
+
+/// Takes the set of attributes of a table:table-cell tag in the ODT's content.xml,
+/// a reference to the map of automatic style names to the map of CSS properties,
+/// and the default cell style name,
+/// then returns a table cell element with styles attached and how many times it should be repeated
+/// (stored inside the element as an attribute called "_repeat")
+fn table_cell_begin(
+    attributes: Attributes,
+    auto_styles: &HashMap<String, HashMap<String, String>>,
+    default_style_name: String,
+) -> Element {
+    let mut style_name = String::new();
+    let mut element = Element::new("td".to_string());
+    for i in attributes {
+        if let Ok(i) = i {
+            let name = std::str::from_utf8(i.key).unwrap_or(":");
+            if name == "table:style-name" {
+                style_name = std::str::from_utf8(
+                    &i.unescaped_value()
+                        .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+                )
+                .unwrap_or("")
+                .to_string();
+            } else if name == "table:number-columns-repeated" {
+                element.attributes.insert(
+                    "_repeat".to_string(),
+                    std::str::from_utf8(
+                        &i.unescaped_value()
+                            .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+                    )
+                    .unwrap_or("1")
+                    .to_string(),
+                );
+            } else if name == "table:number-columns-spanned" {
+                element.attributes.insert(
+                    "colspan".to_string(),
+                    std::str::from_utf8(
+                        &i.unescaped_value()
+                            .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+                    )
+                    .unwrap_or("1")
+                    .to_string(),
+                );
+            } else if name == "table:number-rows-spanned" {
+                element.attributes.insert(
+                    "rowspan".to_string(),
+                    std::str::from_utf8(
+                        &i.unescaped_value()
+                            .unwrap_or_else(|_| std::borrow::Cow::from(vec![])),
+                    )
+                    .unwrap_or("1")
+                    .to_string(),
+                );
+            }
+        }
+    }
+    if default_style_name != "" {
+        element.styles = auto_styles
+            .get(&default_style_name)
+            .unwrap_or(&HashMap::new())
+            .clone();
+    } else {
+        element.styles = auto_styles
+            .get(&style_name)
+            .unwrap_or(&HashMap::new())
+            .clone();
+    }
+    element
+}
